pyproject.toml
[tool.uv]
dev-dependencies = []

[project]
name = "skin-disease-detection"
version = "0.1.0"
description = "AI-powered skin disease detection web application"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "pillow>=10.0.0",
    "httpx>=0.25.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
]
